requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.0
psycopg2-binary
jinja2
pydantic-settings
email-validator
python-multipart
redis
pytest
httpx